dreamsync_repo_v2/dreamsync/requirements.txt
python-dotenv==1.0.1
streamlit==1.36.0
openai==1.35.10
requests==2.32.3
pydub==0.25.1
moviepy==1.0.3
whisper-timestamped==1.15.7
yt_dlp==2024.8.6
ffmpeg-python==0.2.0
tqdm==4.66.4
jinja2==3.1.4